7.4 pre-changes
More code cleanup, favorites pausing, image button tooltips, and
non-sizing change mode.  Modified some process flow to be cleaner and
less wasteful.  Slimmed down code considerably.
