+++ BorderlessGaming.csproj
@@ -91,7 +91,7 @@
     <Compile Include="Forms\MainWindow.Designer.cs">
       <DependentUpon>MainWindow.cs</DependentUpon>
     </Compile>
-    <Compile Include="Favorites.cs" />
+    <Compile Include="Common\Favorites.cs" />
     <Compile Include="Forms\DesktopAreaSelector.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -104,8 +104,8 @@
     <Compile Include="Forms\InputText.designer.cs">
       <DependentUpon>InputText.cs</DependentUpon>
     </Compile>
-    <Compile Include="HiddenProcesses.cs" />
-    <Compile Include="ProcessDetails.cs" />
+    <Compile Include="Common\HiddenProcesses.cs" />
+    <Compile Include="Common\ProcessDetails.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
