+++ Properties/AssemblyInfo.cs
@@ -36,6 +36,6 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: AssemblyVersion("7.4.215.841")]
-[assembly: AssemblyFileVersion("7.4.215.841")]
+[assembly: AssemblyVersion("7.4.215.1285")]
+[assembly: AssemblyFileVersion("7.4.215.1285")]
 [assembly: NeutralResourcesLanguage("en-US")]
